* "Toggle Rerecord Display" hotkey (M), "Play Movie From Beginning" hotkey now also works in Tasedit * Tasedit: input_snapshot also stores "commands" array * Tasedit: Bookmarks/Branches (storing snapshots * Tasedit: Bookmarks list, text flashing * Tasedit: bookmark number images shown in first column of main list * saving/loading Bookmarks in .tas file * Tasedit: Disable F10 as System Key * Tasedit: checkbox Config->Branches restore entire Movie (else restore input only up to and not including bookmarked frame)

+++ src/drivers/win/taseditlib/markers.cpp
@@ -25,7 +25,7 @@ void MARKERS::free()
 
 void MARKERS::update()
 {
-	if (markers_array.size() < currMovieData.getNumRecords())
+	if ((int)markers_array.size() < currMovieData.getNumRecords())
 		markers_array.resize(currMovieData.getNumRecords());
 }
